Moving val modifier processing to new ModiferProcessor
